--- Documentation~/HowToInstall.md
@@ -1,26 +1,0 @@
-# How to install Rosalina
-
-## Installing via OpenUPM
-
-Rosalina is now available via OpenUPM:
-https://openupm.com/packages/com.eastylabs.rosalina/
-
-OpenUPM provides a detailed explanation of how to add packages to unity.
-Just follow the guide in the link:
-![image](install_upm_1.png)
-
-## Installing via git repository
-
-In Unity, navigate to ``Window -> Package Manager``:
-
-![image](install_1.png)
-
-In the ``Package Manager``, click on the ``+`` on the top left and select ``Add package from git URL...``
-
-![image](install_2.png)
-
-No use the following path to install Rosalina ``https://github.com/Eastrall/Rosalina.git``
-
-![image](install_3.png)
-
-You can now start to work with Rosalina.
