+++ Editor/RosalinaGenerationResult.cs.meta
