--- Editor/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: efb8c01b18fb5624aa3daf4c3cece461
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
