--- Editor/Scripts/Parsing/RosalinaUXMLParser.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 00a331a3431761a46a13edff4417d0d4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
