--- Editor/Scripts/RosalinaConstants.cs
@@ -1,6 +1,0 @@
-﻿#if UNITY_EDITOR
-public class RosalinaConstants
-{
-    public const string Version = "1.0.4";
-}
-#endif
