--- Editor/Scripts/RosalinaConstants.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 407c00a38b32ee44db418988709fc344
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
