--- Tools~/Rosalina.Roslyn/Rosalina.Roslyn.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.2.32519.379
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Rosalina.Roslyn", "Rosalina.Roslyn.csproj", "{BC4145A3-EC26-415E-8633-8E453750FE2D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{BC4145A3-EC26-415E-8633-8E453750FE2D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BC4145A3-EC26-415E-8633-8E453750FE2D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BC4145A3-EC26-415E-8633-8E453750FE2D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BC4145A3-EC26-415E-8633-8E453750FE2D}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {35FBCEEB-DA0D-4A0A-9587-B47E0C237746}
-	EndGlobalSection
-EndGlobal
